test(desktop): split mesh command tests
